% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICC.R
\name{AICC}
\alias{AICC}
\title{AICC}
\usage{
AICC(model, n, p)
}
\arguments{
\item{model}{Model to use}

\item{n}{n value}

\item{p}{p value}
}
\value{
Nothing
}
\description{
Compute the AIC for a model
}
\examples{
{
cholesterol.data <- AdvancedRegression::cholesterol

#creating long-form data set
longform.data <- reshape2::melt(cholesterol.data,
                                id.vars = c("id", "gender", "age"),
                                variable.name = "LDLmonth",
                                value.name = "LDL")

#creating numeric variable for time
month <- ifelse(longform.data$LDLmonth == "LDL0", 0,
                ifelse(longform.data$LDLmonth == "LDL6", 6,
                       ifelse(longform.data$LDLmonth == "LDL9", 9, 24)))

#specifying reference category
gender.rel <- relevel(longform.data$gender, ref = "M")

#fitting random slope and intercept model
fitted.model <- nlme::lme(LDL ~ gender.rel + age + month,
                          random =~ 1 + month|id, data = longform.data)
AICC(fitted.model, n = 108, p = 8)
}
}
