% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance}
\alias{deviance}
\title{deviance}
\usage{
deviance(reference_model, fitted_model)
}
\arguments{
\item{reference_model}{The model to act as the reference}

\item{fitted_model}{The model that you want to compare to the reference model}
}
\value{
The deviance between the two models
}
\description{
Given two models, compute the deviance between the two
}
\examples{
{
libraries.data <- AdvancedRegression::libraries

#specifying reference category
location.rel <- relevel(libraries.data$location, ref = "rural")

#fitting beta regression model
fitted.model <- betareg::betareg(propontime ~ nbooks + ncardholders + location.rel,
                                 data = libraries.data, link = "logit")

#checking model fit
intercept.only.model <- betareg::betareg(propontime ~ 1,
                                         data = libraries.data, link = "logit")

deviance(intercept.only.model, fitted.model)
}

}
