% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance_pvalue.R
\name{deviance_pvalue}
\alias{deviance_pvalue}
\title{deviance_pvale}
\usage{
deviance_pvalue(reference_model, fitted_model, df, lower.tail = FALSE)
}
\arguments{
\item{reference_model}{The model to act as the reference}

\item{fitted_model}{The model that you want to compare to the reference model}

\item{df}{degrees of freedom}

\item{lower.tail}{Compute the p-value using the lower tail. Default = FALSE (optional)}
}
\value{
Nothing
}
\description{
Given a reference model and the fitted model and the degrees of freedom
compute the deviance and the p-value. Then print them.
}
\examples{
{
libraries.data <- AdvancedRegression::libraries

#specifying reference category
location.rel <- relevel(libraries.data$location, ref = "rural")

#fitting beta regression model
fitted.model <- betareg::betareg(propontime ~ nbooks + ncardholders + location.rel,
                                 data = libraries.data, link = "logit")

#checking model fit
intercept.only.model <- betareg::betareg(propontime ~ 1,
                                         data = libraries.data, link = "logit")

deviance_pvalue(intercept.only.model, fitted.model, df = 3)
}
}
