% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GammaExample.R
\name{GammaExample}
\alias{GammaExample}
\title{GammaExample}
\usage{
GammaExample()
}
\value{
Nothing
}
\description{
Using the \code{real estate} data, perform a regression using
a box-cox transformation
}
\examples{
{
GammaExample()
}
}
