% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_example.R
\name{get_csv_chapter_example}
\alias{get_csv_chapter_example}
\title{get_csv_chapter_example}
\usage{
get_csv_chapter_example()
}
\description{
Create a tibble with all the chapters and exercise that we have dataset for
}
