% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_col_example.R, R/csv_col_exercise.R
\name{get_csv_cols_example}
\alias{get_csv_cols_example}
\alias{get_csv_cols_chapter_1_example}
\alias{get_csv_cols_chapter_2_example}
\alias{get_csv_cols_chapter_3_example}
\alias{get_csv_cols_chapter_4_example}
\alias{get_csv_cols_chapter_5_example}
\alias{get_csv_cols_chapter_6_example}
\alias{get_csv_cols_chapter_7_example}
\alias{get_csv_cols_chapter_8_example}
\alias{get_csv_cols_chapter_9_example}
\alias{get_csv_cols_chapter_10_example}
\alias{get_csv_cols_exercise}
\alias{get_csv_cols_chapter_1_exercise}
\alias{get_csv_cols_chapter_2_exercise}
\alias{get_csv_cols_chapter_3_exercise}
\alias{get_csv_cols_chapter_4_exercise}
\alias{get_csv_cols_chapter_5_exercise}
\alias{get_csv_cols_chapter_6_exercise}
\alias{get_csv_cols_chapter_7_exercise}
\alias{get_csv_cols_chapter_8_exercise}
\alias{get_csv_cols_chapter_9_exercise}
\alias{get_csv_cols_chapter_10_exercise}
\title{get_csv_cols_example}
\usage{
get_csv_cols_example(chapter, exercise)

get_csv_cols_chapter_1_example(exercise)

get_csv_cols_chapter_2_example(exercise)

get_csv_cols_chapter_3_example(exercise)

get_csv_cols_chapter_4_example(exercise)

get_csv_cols_chapter_5_example(exercise)

get_csv_cols_chapter_6_example(exercise)

get_csv_cols_chapter_7_example(exercise)

get_csv_cols_chapter_8_example(exercise)

get_csv_cols_chapter_9_example(exercise)

get_csv_cols_chapter_10_example(exercise)

get_csv_cols_exercise(chapter, exercise)

get_csv_cols_chapter_1_exercise(exercise)

get_csv_cols_chapter_2_exercise(exercise)

get_csv_cols_chapter_3_exercise(exercise)

get_csv_cols_chapter_4_exercise(exercise)

get_csv_cols_chapter_5_exercise(exercise)

get_csv_cols_chapter_6_exercise(exercise)

get_csv_cols_chapter_7_exercise(exercise)

get_csv_cols_chapter_8_exercise(exercise)

get_csv_cols_chapter_9_exercise(exercise)

get_csv_cols_chapter_10_exercise(exercise)
}
\arguments{
\item{chapter}{Book chapter}

\item{exercise}{Exercise number}
}
\description{
Get the col() information for reading CSV files
}
