% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_example.R
\name{get_csv_filepath_example}
\alias{get_csv_filepath_example}
\title{get_csv_filepath_example}
\usage{
get_csv_filepath_example(chapter, exercise, path_base = file.path(".",
  "inst", "example"))
}
\arguments{
\item{chapter}{Chapter to create the file path for}

\item{exercise}{Exercise to create the file path for}

\item{path_base}{Base path}
}
\description{
create the filepath for the raw CSV file
}
