% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_exercise.R
\name{get_csv_filepath_execise}
\alias{get_csv_filepath_execise}
\title{get_csv_filepath_execise}
\usage{
get_csv_filepath_execise(chapter, exercise, path_base = file.path(".",
  "inst", "exercise"))
}
\arguments{
\item{chapter}{Chapter to create the file path for}

\item{exercise}{Exercise to create the file path for}

\item{path_base}{Base path}
}
\description{
create the filepath for the raw CSV file
}
