% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_na_example.R, R/csv_na_exercise.R
\name{get_csv_nas_example}
\alias{get_csv_nas_example}
\alias{get_csv_nas_exercise}
\title{get_csv_nas}
\usage{
get_csv_nas_example(chapter, exercise)

get_csv_nas_exercise(chapter, exercise)
}
\arguments{
\item{chapter}{Book chapter}

\item{exercise}{Exercise number}
}
\description{
Get the 'NA' for reading CSV files
}
