% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalExample.R
\name{HierarchialExample}
\alias{HierarchialExample}
\title{HierachialExample}
\usage{
HierarchialExample()
}
\value{
Nothing
}
\description{
Creates a Hierachial model using family depression data \code{dyads} and
prints the results
}
\examples{
{
HierarchialExample()
}
}
