% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_example.R
\name{load_csv_example}
\alias{load_csv_example}
\title{load_csv_example}
\usage{
load_csv_example(chapter, exercise)
}
\arguments{
\item{chapter}{Chapter to load}

\item{exercise}{Exercise in the chapter to load}
}
\value{
Nothing
}
\description{
Load the CSV file
}
