% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticExample.R
\name{LogisticExample}
\alias{LogisticExample}
\title{LogisticExample()}
\usage{
LogisticExample()
}
\value{
Nothing
}
\description{
Create a logistic model using the \code{companies} data.
}
\examples{
{
LogisticExample()
}
}
