% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalExample_1.R
\name{LongitudinalExample_1}
\alias{LongitudinalExample_1}
\title{LongitudinalExample_1}
\usage{
LongitudinalExample_1()
}
\value{
Nothing
}
\description{
Create a repeated measures regression using the \code{cholesterol} data
}
\examples{
{
LongitudinalExample_1()
}
}
