% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongitudinalExample_2.R
\name{LongitudinalExample_2}
\alias{LongitudinalExample_2}
\title{LongitudinalExample_2}
\usage{
LongitudinalExample_2()
}
\value{
Nothing
}
\description{
Create a repeated measures regression using the \code{cholesterol} data
}
\examples{
{
LongitudinalExample_2()
}
}
