% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_example.R
\name{make_data_example}
\alias{make_data_example}
\title{make_data_example}
\usage{
make_data_example(df = get_csv_chapter_example())
}
\arguments{
\item{df}{data.frame to use}
}
\description{
Batch make the RDA files
}
