% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_exercise.R
\name{make_data_exercise}
\alias{make_data_exercise}
\title{make_data_exercise}
\usage{
make_data_exercise(df = get_csv_chapter_exercise())
}
\arguments{
\item{df}{data.frame to use}
}
\description{
Batch make the RDA files
}
