% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalExample.R
\name{NormalExample}
\alias{NormalExample}
\title{NormalExample}
\usage{
NormalExample()
}
\value{
Nothing
}
\description{
Create a linear regression based on \code{job satisfaction} data
}
\examples{
{
NormalExample()
}
}
