% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoissonExample.R
\name{PoissonExample}
\alias{PoissonExample}
\title{PoissonExample}
\usage{
PoissonExample()
}
\value{
Nothing
}
\description{
Poisson regression using \code{hospital stay} data
}
\examples{
{
PoissonExample()
}
}
