% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_prediction.R
\name{print_prediction}
\alias{print_prediction}
\title{print_prediction}
\usage{
print_prediction(model, predict_data, ...)
}
\arguments{
\item{model}{Model used to make the prediction}

\item{predict_data}{Data used to make the prediciton}

\item{...}{Any extra parameters to be passed the \code{predict}}
}
\value{
Nothing
}
\description{
Given a model and data, print the predicted value.
}
\examples{
{
job.satisfaction.data <- AdvancedRegression::job_satisfaction
#specifying reference levels
educ.rel <- relevel(job.satisfaction.data$educ, ref = "master")
gender.rel <- relevel(job.satisfaction.data$gender, ref = "F")

#fitting general linear model
fitted.model <- glm(score ~ gender.rel + age + educ.rel,
                    data = job.satisfaction.data,
                    family = gaussian(link = identity))
#using fitted model for prediction
print_prediction(fitted.model,
                 tibble::tibble(gender.rel = "F", age = 40, educ.rel = "bachelor"))
}
}
