% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_example.R
\name{save_data_example}
\alias{save_data_example}
\title{save_data_example
Create a RDA file for a chapter and exercise}
\usage{
save_data_example(chapter, exercise)
}
\arguments{
\item{chapter}{Example chapter number}

\item{exercise}{Example exercise number}
}
\description{
save_data_example
Create a RDA file for a chapter and exercise
}
