% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_exercise.R
\name{save_data_exercise}
\alias{save_data_exercise}
\title{save_data_exercise
Create a RDA file for a chapter and exercise}
\usage{
save_data_exercise(chapter, exercise)
}
\arguments{
\item{chapter}{Example chapter number}

\item{exercise}{Example exercise number}
}
\description{
save_data_exercise
Create a RDA file for a chapter and exercise
}
