% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPExample.R
\name{ZIPExample}
\alias{ZIPExample}
\title{ZIPExample}
\usage{
ZIPExample()
}
\value{
Nothing
}
\description{
Zero inflated Poison regression using \code{smoking} data with
the \code{pscl} package
}
\examples{
{
ZIPExample()
}
}
