% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerciseData.R
\docType{data}
\name{BetaExerciseData}
\alias{BetaExerciseData}
\alias{GammaExerciseData}
\alias{HierarchicalExerciseData}
\alias{LogisticExerciseData}
\alias{LongitudinalExerciseData}
\alias{NormalExerciseData}
\alias{PoissonExerciseData}
\alias{ZIPExerciseData}
\title{Exercise Data}
\format{An object of class \code{data.frame} with 20 rows and 5 columns.}
\usage{
BetaExerciseData

GammaExerciseData

HierarchicalExerciseData

LogisticExerciseData

LongitudinalExerciseData

NormalExerciseData

PoissonExerciseData

ZIPExerciseData
}
\description{
Data used for the different exercises
}
\keyword{datasets}
