% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_col_example.R, R/csv_col_exercise.R
\name{get_csv_cols_example}
\alias{get_csv_cols_example}
\alias{get_csv_cols_exercise}
\title{get_csv_cols_example}
\usage{
get_csv_cols_example(chapter, exercise)

get_csv_cols_exercise(chapter, exercise)
}
\arguments{
\item{chapter}{Book chapter}

\item{exercise}{Exercise number}
}
\description{
Get the col() information for reading CSV files
}
